\name{aes_default}
\alias{aes_default}
\title{Default aesthetics.}
\usage{
  aes_default(geom)
}
\value{
  named list of default values
}
\description{
  These values are used to fill in missing aesthetics to
  provide default values. See \code{\link{build_defaults}}
  for a way to build defaults from a common list of
  options.
}

