\name{aes_required}
\alias{aes_required}
\title{Required aesthetics.}
\usage{
  aes_required(geom)
}
\value{
  character vector of aesthetic names
}
\description{
  Required aesthetics must be present, otherwise the geom
  can not be drawn. These are usually position aesthetics,
  as all others are supplied by the defaults.
}

