% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layer_class.R
\name{expected}
\alias{expected}
\alias{expected.layer}
\title{Compute the expected losses ceded to the layer.}
\usage{
expected(layer)

\method{expected}{layer}(layer)
}
\description{
Compute the expected losses ceded to the layer.
}
\examples{
expected(test_layer)
}
