% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/portfolio.R
\name{explode}
\alias{explode}
\title{Explodes a list of porfolios into a list of layers.}
\usage{
explode(L)
}
\arguments{
\item{L}{the layers to be exploded}
}
\value{
A list containing all the layers in L
}
\description{
Explodes a list of porfolios into a list of layers.
}
