\name{geom_deparse}
\alias{geom_deparse}
\title{Deparse a geom into the call that created it.
Useful for serialising ggplot2 objects back into R code.}
\usage{
  geom_deparse(geom)
}
\description{
  Deparse a geom into the call that created it. Useful for
  serialising ggplot2 objects back into R code.
}

