\name{geom_plot}
\alias{geom_plot}
\title{Convenience method for plotting geoms.}
\usage{
  geom_plot(geom, data = list(), munch = FALSE)
}
\description{
  This illustrates the basic pipeline by which the geom
  methods are called, and makes it easier to visually test
  geoms, but graphics packages will usually call the
  methods individually themselves.
}

