\name{geom_range}
\alias{geom_range}
\title{Compute the range of the data supplied to a geom.}
\usage{
  geom_range(geom, data)
}
\description{
  The default method inspects x and y aesthetics and data
  values. This function should be called with data passed
  through \code{\link{geom_data}} (and possibly through
  \code{\link{geom_premunch}}) so only the data component
  generally needs to be inspected.
}

