\name{geom_stat}
\alias{geom_adjust}
\alias{geom_stat}
\title{Default related stat and adjust.}
\usage{
  geom_stat(geom, ...)
}
\arguments{
  \item{...}{Other arguments passed on to the object
  creation function.  These are used when initialised a
  geom from a name and list of parameters, which may belong
  to the stat or the position adjustment.}
}
\description{
  The defaults are the identity functions which leave the
  data unchanged.
}

