\name{join_aesthetics}
\alias{join_aesthetics}
\title{Join constant aesthetics back onto results of statistical transformation.}
\usage{
  join_aesthetics(stats, data)
}
\description{
  This function ensures that all group-wise constant
  aesthetics are preserved after statistical
  transformation.
}

