% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layer_class.R
\name{metrics}
\alias{metrics}
\alias{metrics.layer}
\title{Compute a list of metrics for the layer.}
\usage{
metrics(layer)

\method{metrics}{layer}(layer)
}
\arguments{
\item{layer}{The layer to calculate metrics for.}
}
\value{
An object of class metric_list containing mean, standard deviation, VaR and tVaR (AEP).
}
\description{
Compute a list of metrics for the layer.
}
\examples{
layer(4000000, 1000000, 1, "yelt", c("GL", "AUTO"), 0, UNLIMITED)
}
