% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layer_class.R
\name{print.layer}
\alias{print.layer}
\title{Print function for objects of class layer.}
\usage{
\method{print}{layer}(x, ...)
}
\arguments{
\item{x}{The layer to be printed.}

\item{...}{Objects to be passed to subsequent methods, if they existed.}
}
\description{
Print function for objects of class layer.
}
\examples{
test_layer <- layer(4000000, 1000000, 1, "yelt_test", lobs=c("PHYSICIANS","CHC","MEDCHOICE"))
test_layer
print(test_layer)
}
