% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layer_class.R, R/old_layer_class.R
\name{print.metric_list}
\alias{print.metric_list}
\title{Print function for objects of class metric_list}
\usage{
\method{print}{metric_list}(x)

\method{print}{metric_list}(x)
}
\description{
Print function for objects of class metric_list

Print function for objects of class metric_list
}
\examples{
metrics(example_layer)
print(metrics(example_layer))
metrics(example_layer)
print(metrics(example_layer))
}
