% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/portfolio.R
\name{print.portfolio}
\alias{print.portfolio}
\title{Print function for objects of class portfolio.}
\usage{
\method{print}{portfolio}(x, ...)
}
\arguments{
\item{x}{The layer to be printed.}

\item{...}{Objects to be passed to subsequent methods, if they existed.}
}
\description{
Print function for objects of class portfolio.
}
\examples{
layer1 <- layer(4000000, 1000000, 1, "yelt_test", lobs="PHYSICIANS")
layer2 <- layer(5000000, 5000000, 1, "yelt_test", lobs="PHYSICIANS")
P <- portfolio(layer1, layer2)
P
print(P)
}
