% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layer.R
\name{print.summary.layer}
\alias{print.summary.layer}
\title{Print function for objects of class summary.layer}
\usage{
\method{print}{summary.layer}(x, ...)
}
\arguments{
\item{x}{The summary to be printed.}

\item{...}{Objects to be passed to subsequent methods, if they existed.}
}
\description{
Print function for objects of class summary.layer
}
\examples{
test_layer <- layer(4000000, 1000000, 1, "yelt_test", lobs=c("PHYSICIANS","CHC","MEDCHOICE"))
summary(test_layer)
print(summary(test_layer)) # same thing
}
