% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/portfolio.R
\name{print.summary.portfolio}
\alias{print.summary.portfolio}
\title{Print function for objects of class summary.portfolio}
\usage{
\method{print}{summary.portfolio}(x, ...)
}
\arguments{
\item{x}{The summary to be printed.}

\item{...}{Objects to be passed to subsequent methods, if they existed.}
}
\description{
Print function for objects of class summary.portfolio
}
\examples{
layer1 <- layer(100000, 100000, 1, "yelt_test", lobs="PHYSICIANS")
layer2 <- layer(100000, 200000, 1, "yelt_test", lobs="PHYSICIANS")
layer3 <- layer(100000, 300000, 1, "yelt_test", lobs="PHYSICIANS")
P <- portfolio(layer1, layer2, layer3)
summary(P)
print(summary(P)) # same thing
}
