\name{stat_densitypoint}
\alias{stat_densitypoint}
\title{Compute the density at each point in the data}
\usage{
  stat_densitypoint(scale = TRUE)
}
\arguments{
  \item{scale}{should x and y positions be scaled
  independently?}
}
\description{
  Requires the locfit package.
}
\seealso{
  \code{\link{locfit}} for the algorithm used to calculate
  the density
}

