\name{stat_transform.densitypoint}
\alias{stat_transform.densitypoint}
\title{data(diamonds, package = "ggplot2")
df <- diamonds[c("price", "carat")]
names(df) <- c("x", "y")
stat_transform(stat_densitypoint(), df)}
\usage{
  stat_transform.densitypoint(stat, data, xrange, yrange)
}
\description{
  data(diamonds, package = "ggplot2") df <-
  diamonds[c("price", "carat")] names(df) <- c("x", "y")
  stat_transform(stat_densitypoint(), df)
}

