% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layer_class.R
\name{stdev}
\alias{stdev}
\alias{stdev.layer}
\alias{stdev.portfolio}
\title{Compute the standard deviation of losses ceded to the layer.}
\usage{
stdev(object)

\method{stdev}{layer}(object)

\method{stdev}{portfolio}(object)
}
\arguments{
\item{object}{the layer or portfolio to compute the standard deviation of}
}
\description{
Compute the standard deviation of losses ceded to the layer.
}
\examples{
test_layer <- layer(4000000, 1000000, 1, "yelt_test", lobs=c("PHYSICIANS","CHC","MEDCHOICE"))
stdev(test_layer)
}
