% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layer_class.R
\name{summary.layer}
\alias{summary.layer}
\title{Summarize the layer parameters, and compute some metrics
for the layer.}
\usage{
\method{summary}{layer}(layer)
}
\arguments{
\item{layer}{The layer to calculate metrics for.}
}
\value{
An object of class summary.layer containing layer parameters, mean,
standard deviation, VaR and tVaR (AEP).
}
\description{
Summarize the layer parameters, and compute some metrics
for the layer.
}
\examples{
test_layer <- layer(4000000, 1000000, 1, "yelt_test", lobs=c("PHYSICIANS","CHC","MEDCHOICE"))
summary(test_layer)
}
