% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/portfolio.R
\name{summary.portfolio}
\alias{summary.portfolio}
\title{Summarize the portfolio parameters, and compute some metrics
for the portfolio}
\usage{
\method{summary}{portfolio}(object, ...)
}
\arguments{
\item{object}{The portfolio to calculate metrics for.}

\item{...}{Objects to be passed to subsequent methods, if they existed.}
}
\value{
An object of class summary.portfolio containing layer parameters, mean,
standard deviation, VaR and tVaR (AEP).
}
\description{
Summarize the portfolio parameters, and compute some metrics
for the portfolio
}
\examples{
layer1 <- layer(100000, 100000, 1, "yelt_test", lobs="PHYSICIANS")
layer2 <- layer(100000, 200000, 1, "yelt_test", lobs="PHYSICIANS")
layer3 <- layer(100000, 300000, 1, "yelt_test", lobs="PHYSICIANS")
P <- portfolio(layer1, layer2, layer3)
summary(P)
}
