% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layer_class.R
\name{tVaR}
\alias{tVaR}
\alias{tVaR.layer}
\title{Compute tail value at risk for the losses in the layer.}
\usage{
tVaR(layer, q)

\method{tVaR}{layer}(layer, q)
}
\arguments{
\item{layer}{the layer to computer tVaR with.}

\item{q}{Quantile for tVaR. For example, if the return period is 100 years, q = 1 - 1/100.}
}
\description{
Compute tail value at risk for the losses in the layer.
}
\examples{
test_layer <- layer(4000000, 1000000, 1, "yelt_test", lobs=c("PHYSICIANS","CHC","MEDCHOICE"))
tVaR(test_layer, 1 - 1/25)
}
