% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layer_class.R, R/old_layer_class.R
\docType{data}
\name{UNLIMITED}
\alias{UNLIMITED}
\title{Convenient synonym for .Machine$double.xmax}
\format{
An object of class \code{numeric} of length 1.

An object of class \code{numeric} of length 1.
}
\usage{
UNLIMITED

UNLIMITED
}
\description{
Convenient synonym for .Machine$double.xmax

Convenient synonym for .Machine$double.xmax
}
\keyword{datasets}
