% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layer_class.R
\name{VaR}
\alias{VaR}
\alias{VaR.layer}
\title{Compute value at risk for the losses in the layer.}
\usage{
VaR(layer, ...)

\method{VaR}{layer}(layer, q, type = c("AEP", "OEP"))
}
\description{
Compute value at risk for the losses in the layer.
}
\examples{
VaR(test_layer, 1 - 1/25)
}
