% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global.R
\name{AAL}
\alias{AAL}
\title{A function to calculate Average Annual Loss (AAL) from a yelt}
\usage{
AAL(yelt)
}
\arguments{
\item{yelt}{The yelt as a data frame or tibble. It must contain columns named trialID, LOB, and Loss}
}
\description{
A function to calculate Average Annual Loss (AAL) from a yelt
}
\examples{
AAL(yelt_test)
}
