% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layer_class.R
\name{expected}
\alias{expected}
\alias{expected.layer}
\title{Compute the expected losses ceded to the layer.}
\usage{
expected(layer)

\method{expected}{layer}(layer)
}
\arguments{
\item{layer}{the layer to compute the expectation of}
}
\description{
Compute the expected losses ceded to the layer.
}
\examples{
test_layer <- layer(4000000, 1000000, 1, "yelt_test", lobs=c("PHYSICIANS","CHC","MEDCHOICE"))
expected(test_layer)
}
