% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layer.R
\name{layer}
\alias{layer}
\title{Create a layer object.}
\usage{
layer(
  limit,
  attachment,
  participation = 1,
  loss_set,
  lobs,
  agg_attachment = 0,
  agg_limit = UNLIMITED
)
}
\arguments{
\item{limit}{layer per occurrence limit.}

\item{attachment}{layer per occurrence attachment point.}

\item{participation}{layer participation (proportion ceded).}

\item{loss_set}{String containing the name of data frame. The dataframe must contain columns trialID, LOB, and Loss.}

\item{lobs}{list of LOBs the layer applies to.}

\item{agg_attachment}{layer per occurrence limit.}

\item{agg_limit}{layer per occurrence limit.}
}
\value{
The layer object.
}
\description{
Create a layer object.
}
\examples{
test_layer <- layer(4000000, 1000000, 1, "yelt_test", lobs=c("PHYSICIANS","CHC","MEDCHOICE"))
agg_layer <- layer(4000000, 1000000, 1, "yelt_test", lobs=c("PHYSICIANS","CHC","MEDCHOICE"),
                   agg_attachment = 4000000, agg_limit=8000000)
}
