% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global.R, R/layer.R, R/portfolio.R
\name{minus}
\alias{minus}
\alias{minus.layer}
\alias{minus.portfolio}
\title{Change the sign of the losses in a layer.}
\usage{
minus(object)

\method{minus}{layer}(object)

\method{minus}{portfolio}(object)
}
\arguments{
\item{object}{the layer to change the sign of}
}
\description{
Change the sign of the losses in a layer.
}
\examples{
test_layer <- layer(4000000, 1000000, 1, "yelt_test", lobs=c("PHYSICIANS","CHC","MEDCHOICE"))
minus(test_layer)
}
