% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layer_class.R
\name{portfolio}
\alias{portfolio}
\title{Create a portfolio object.}
\usage{
portfolio(layer_list)
}
\arguments{
\item{layer_list}{a list of layer objects}
}
\value{
The portfolio object.
}
\description{
Create a portfolio object.
}
\examples{
portfolio(list(layer1, layer2, layer3))
}
