% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/portfolio.R
\name{portfolio}
\alias{portfolio}
\title{Create a portfolio object.}
\usage{
portfolio(...)
}
\arguments{
\item{...}{layers in the portfolio}
}
\value{
The portfolio object.
}
\description{
Create a portfolio object.
}
