% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layer_class.R, R/old_layer_class.R
\name{print.layer}
\alias{print.layer}
\title{Print function for objects of class layer.}
\usage{
\method{print}{layer}(layer)

\method{print}{layer}(layer)
}
\description{
Print function for objects of class layer.

Print function for objects of class layer.
}
\examples{
test_layer
print(test_layer)
test_layer
print(test_layer)
}
