% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layer_class.R
\name{stdev}
\alias{stdev}
\alias{stdev.layer}
\title{Compute the standard deviation of losses ceded to the layer.}
\usage{
stdev(layer)

\method{stdev}{layer}(layer)
}
\description{
Compute the standard deviation of losses ceded to the layer.
}
\examples{
stdev(test_layer)
}
