% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layer_class.R
\name{tVaR}
\alias{tVaR}
\alias{tVaR.layer}
\title{Compute tail value at risk for the losses in the layer.}
\usage{
tVaR(layer, ...)

\method{tVaR}{layer}(layer, q)
}
\description{
Compute tail value at risk for the losses in the layer.
}
\examples{
test_layer <- layer(4000000, 1000000, 1, "yelt_test", lobs=c("PHYSICIANS","CHC","MEDCHOICE"))
tVaR(test_layer, 1 - 1/25)
}
