% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global.R, R/layer_class.R, R/portfolio.R
\name{tVaR}
\alias{tVaR}
\alias{tVaR.layer}
\alias{tVaR.portfolio}
\title{Compute tail value at risk for the losses in the layer.}
\usage{
tVaR(object, rp_years, type = c("AEP", "OEP"))

\method{tVaR}{layer}(object, rp_years, type = c("AEP", "OEP"))

\method{tVaR}{portfolio}(object, rp_years, type = c("AEP", "OEP"))
}
\arguments{
\item{object}{the layer or portfolio to computer VaR with.}

\item{rp_years}{Number of years in the return period}

\item{type}{AEP (aggregate exceedance probability)or OEP (occurrence exceedance probability). Defaults to AEP.}
}
\description{
Compute tail value at risk for the losses in the layer.
}
\examples{
gross_layer <- layer(UNLIMITED, 0, 1, "yelt_test", lobs=c("PHYSICIANS","CHC","MEDCHOICE"))
tVaR(gross_layer, 25)
tVaR(gross_layer, 25, "AEP") # the same thing
tVaR(gross_layer, 25, "OEP")
}
