% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global.R
\docType{data}
\name{UNLIMITED}
\alias{UNLIMITED}
\title{A convenient synonym for .Machine$double.xmax; used for the limit parameter when there is
no limit.}
\format{
An object of class \code{numeric} of length 1.
}
\usage{
UNLIMITED
}
\description{
A convenient synonym for .Machine$double.xmax; used for the limit parameter when there is
no limit.
}
\keyword{datasets}
