% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layer_class.R
\name{VaR}
\alias{VaR}
\alias{VaR.layer}
\alias{VaR.portfolio}
\title{Compute value at risk for the losses in the layer.}
\usage{
VaR(object, rp_years, type = c("AEP", "OEP"))

\method{VaR}{layer}(object, rp_years, type = c("AEP", "OEP"))

\method{VaR}{portfolio}(object, rp_years, type = c("AEP", "OEP"))
}
\arguments{
\item{object}{the layer or portfolio to computer VaR with.}

\item{rp_years}{Number of years in the return period}

\item{type}{AEP (aggregate exceedance probability)or OEP (occurrence exceedance probability). Defaults to AEP.}
}
\description{
Compute value at risk for the losses in the layer.
}
\examples{
gross_layer <- layer(UNLIMITED, 0, 1, "yelt_test", lobs=c("PHYSICIANS","CHC","MEDCHOICE"))
VaR(gross_layer, 25)
VaR(gross_layer, 25, "AEP") # the same thing
VaR(gross_layer, 25, "OEP")
}
