% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layer_class.R
\name{VaR}
\alias{VaR}
\alias{VaR.layer}
\title{Compute value at risk for the losses in the layer.}
\usage{
VaR(layer, ...)

\method{VaR}{layer}(layer, q, type = c("AEP", "OEP"))
}
\arguments{
\item{layer}{the layer to computer VaR with.}

\item{q}{Quantile for VaR. For example, if the return period is 100 years, q = 1 - 1/100.}

\item{type}{AEP (aggregate exceedance probability)or OEP (occurrence exceedance probability). Defaults to AEP.}
}
\description{
Compute value at risk for the losses in the layer.
}
\examples{
test_layer <- layer(4000000, 1000000, 1, "yelt_test", lobs=c("PHYSICIANS","CHC","MEDCHOICE"))
VaR(test_layer, 1 - 1/25)
VaR(test_layer, 1 - 1/25, "AEP") # the same thing
VaR(test_layer, 1 - 1/25, "OEP")
}
