% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layer_class.R
\docType{data}
\name{yelt_test}
\alias{yelt_test}
\title{A sample YELT (year event loss table) to test layers functions on}
\format{
A data frame with 127648 rows and 5 variables:
\describe{
\item{trialID}{Identifier for the trial, which represents one year}
\item{LOB}{Line of business or other segmentation}
\item{Loss}{Loss amount}
\item{Sequence}{Sequence of loss within trial}

}
}
\usage{
yelt_test
}
\description{
A sample YELT (year event loss table) to test layers functions on
}
\keyword{datasets}
